Address provider search review comments
Keep the provider search live region mounted for reliable announcements and move repeated provider test entries into the shared fixture.
